requirements.txt
splunk-sdk>=1.6.0
python-dotenv>=1.0.0
